requirements.txt
google-cloud-pubsub
apache-beam
apache-beam[gcp]
wheel
python3.7
